+++ actiontext/test/dummy/app/models/review.rb
@@ -2,4 +2,5 @@ class Review < ApplicationRecord
   belongs_to :message
 
   has_rich_text :content
+  has_rich_text :rich_content, column: true
 end