+++ actiontext/test/dummy/db/migrate/20230824140534_add_rich_content_to_messages.rb
@@ -1,0 +1,7 @@
+class AddRichContentToMessages < ActiveRecord::Migration[6.0]
+  def change
+    change_table :messages do |t|
+      t.text :rich_content
+    end
+  end
+end