+++ actiontext/test/dummy/db/migrate/20230824141226_add_rich_content_to_reviews.rb
@@ -1,0 +1,7 @@
+class AddRichContentToReviews < ActiveRecord::Migration[6.0]
+  def change
+    change_table :reviews do |t|
+      t.text :rich_content
+    end
+  end
+end