+++ actiontext/test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2019_03_17_200724) do
+ActiveRecord::Schema[7.1].define(version: 2023_08_24_141226) do
   create_table "action_text_rich_texts", force: :cascade do |t|
     t.string "name", null: false
     t.text "body"
@@ -53,6 +53,7 @@ ActiveRecord::Schema[7.1].define(version: 2019_03_17_200724) do
     t.string "subject"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.string "rich_content"
   end
 
   create_table "pages", force: :cascade do |t|
@@ -70,6 +71,7 @@ ActiveRecord::Schema[7.1].define(version: 2019_03_17_200724) do
   create_table "reviews", force: :cascade do |t|
     t.integer "message_id", null: false
     t.string "author_name", null: false
+    t.string "rich_content"
     t.index ["message_id"], name: "index_reviews_on_message_id"
   end